refactor: collapse 11 identical GeorgianVerbsQuizCommandN into one parameterised class
GeorgianVerbsQuizCommand1..11 were 842 lines of near-identical code differing
only in the lesson number (1–11). Replaced with GeorgianVerbsQuizStartCommand
that takes lessonId in its primary constructor; DI registration loops 1..11
using a factory lambda. FEATURES.md entry updated to reflect the new class name.

All 1402 tests green, 13 Playwright specs pass.
